+++ R/yesno2.R
@@ -1,27 +1,34 @@
 #' Yes No with Two Options
 #'
-#' Asks a custom yes-no question with 'Yes' or 'No' (in that order) as the only responses.
+#' Asks a custom yes-no question with two responses (by default 'Yes' or 'No').
 #' Returns a flag indicating which response the user choose.
-#' It is designed to be used in situations where a user needs to choose one of two options both of which have consequences.
+#' It is designed to be used in situations where a user needs to choose
+#' one of two affirmative options.
 #'
 #' The objects are first pasted without separators
 #' and collapsed using \code{\link{paste0}(..., collapse = "")}
 #' before being output using \code{\link{cat}}.
 #'
-#' The order and phrasing is fixed so that users can quickly type their
-#' preferred response to the question.
-#'
-#' Selection of a 0 (to exit) is also considered to be No.
+#' Selection of a 0 (to exit) causes the code to throw an error.
 #'
 #' @param ... Objects to paste and then output to produce the question.
-#' @return A flag indicating whether the user answered yes or no.
+#' @param yes A string of the first response.
+#' @param no A string of the second response.
+#' @return A flag indicating whether the user selected the first (TRUE) or second (FALSE) response.
 #' @seealso \code{\link{yesno}}
 #' @export
 #' @examples
 #'\dontrun{
-#' yesno2("Do you like ", R.Version()$nickname ,"?")
+#' yesno2("Do you like this question?", yes = "I really do")
 #'}
-yesno2 <- function(...) {
+yesno2 <- function(..., yes = "Yes", no = "No") {
+  if(!(is.character(yes) && identical(length(yes), 1L)) && isTRUE(!is.na(yes)))
+    stop("yes must be a string")
+  if(!(is.character(no) && identical(length(no), 1L)) && isTRUE(!is.na(no)))
+    stop("no must be a string")
+
   cat(paste0(..., collapse = ""))
-  menu(c("Yes", "No")) == 1
+  response <- menu(c(yes, no))
+  if (response == 0) stop("user choose to exit")
+  response == 1
 }